Long类型比较相等问题调整

+++ ruoyi-admin/src/main/java/com/ruoyi/web/controller/system/SysUserController.java
@@ -192,7 +192,7 @@ public class SysUserController extends BaseController
         user.setPassword(passwordService.encryptPassword(user.getLoginName(), user.getPassword(), user.getSalt()));
         if (userService.resetUserPwd(user) > 0)
         {
-            if (ShiroUtils.getUserId() == user.getUserId())
+            if (ShiroUtils.getUserId().longValue() == user.getUserId().longValue())
             {
                 ShiroUtils.setSysUser(userService.selectUserById(user.getUserId()));
             }
